requirements.txt
pydantic>=2.0.0
pyyaml>=6.0
httpx>=0.24.0
asyncpg>=0.28.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
sqlparse>=0.4.4
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
tenacity>=8.2.0
regex>=2023.10.3
respx>=0.20.0
deepeval>=1.0.0
